# docs/openapi.json
{
 "components": {
  "schemas": {
   "DeviceConfigOut": {
    "properties": {
     "clamp_notes": {
      "default": [],
      "items": {
       "type": "string"
      },
      "title": "Clamp Notes",
      "type": "array"
     },
     "config": {
      "additionalProperties": true,
      "title": "Config",
      "type": "object"
     },
     "config_version": {
      "title": "Config Version",
      "type": "string"
     },
     "device_id": {
      "title": "Device Id",
      "type": "string"
     },
     "is_default": {
      "title": "Is Default",
      "type": "boolean"
     },
     "publish_warning": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Publish Warning"
     },
     "published_to": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Published To"
     },
     "site_id": {
      "title": "Site Id",
      "type": "string"
     },
     "updated_utc": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated Utc"
     },
     "version": {
      "title": "Version",
      "type": "integer"
     }
    },
    "required": [
     "device_id",
     "site_id",
     "version",
     "config_version",
     "is_default",
     "updated_utc",
     "config"
    ],
    "title": "DeviceConfigOut",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "OceanKind",
  "version": "2.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/admin/devices/{device_pk}": {
   "delete": {
    "description": "Revocation. The unit gets 401 on its next poll and keeps its last valid\nconfig, per DATA-CONTRACT.md expiry semantics.",
    "operationId": "delete_device_api_admin_devices__device_pk__delete",
    "parameters": [
     {
      "in": "path",
      "name": "device_pk",
      "required": true,
      "schema": {
       "title": "Device Pk",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Device",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/admin/devices/{device_pk}/config": {
   "get": {
    "operationId": "get_device_config_api_admin_devices__device_pk__config_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_pk",
      "required": true,
      "schema": {
       "title": "Device Pk",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceConfigOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Device Config",
    "tags": [
     "admin"
    ]
   },
   "put": {
    "description": "Tune, then publish.\n\nFull replace. Missing fields take defaults; unknown fields are an error,\nbecause a typo'd key that silently tuned nothing is a quiet failure.\n\nThe saved values are then written to `sites/{site_id}/remote_config.json`,\nsigned. That blob is the delivery path (D-020) \u2014 saving without publishing\nwould leave the panel showing a configuration no device will ever apply.",
    "operationId": "put_device_config_api_admin_devices__device_pk__config_put",
    "parameters": [
     {
      "in": "path",
      "name": "device_pk",
      "required": true,
      "schema": {
       "title": "Device Pk",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "additionalProperties": true,
        "title": "Body",
        "type": "object"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceConfigOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Device Config",
    "tags": [
     "admin"
    ]
   }
  },
...
  "/api/devices/config": {
   "get": {
    "description": "**Read-only debugging view of the configuration blob** (D-020).\n\nThis is not the delivery path. The device reads\n`sites/{site_id}/remote_config.json` from storage, which the backend writes\nwhen configuration is tuned. This route exists so an operator can see what a\nunit will receive without opening a storage browser, and it returns the blob\n**byte for byte** \u2014 reading it here and reading it from storage must never\nproduce two different documents, because that is how a signature mismatch\nhides.\n\nIt deliberately does not compose a document of its own. If the blob is\nabsent, that is the honest answer: nothing has been published for this site.",
    "operationId": "device_config_api_devices_config_get",
    "parameters": [
     {
      "in": "header",
      "name": "x-device-key",
      "required": true,
      "schema": {
       "title": "X-Device-Key",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-device-id",
      "required": true,
      "schema": {
       "title": "X-Device-Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Device Config",
    "tags": [
     "devices"
    ]
   }
  },
  "/api/devices/events": {
   "post": {
    "description": "**One detection, pushed by the device as it happens** (R-6.3, D-022).\n\nThis is the low-latency path and nothing more. The device writes the event\nblob regardless of what happens here, so a failure costs freshness and never\nan event: the weekly reconcile pass reads the same event out of storage. Any\nchange that makes correctness depend on this route has broken the design \u2014\nsee `DATA-CONTRACT.md`, **Event upload**.\n\n**`202` whether or not the event was already indexed.** A duplicate is the\nexpected case, not an error: a reconcile pass may have picked the event up\nfirst, or the device may be retrying after an ambiguous timeout. Returning\n`409` would push the device into tracking what it had successfully sent,\nwhich is exactly the bookkeeping the unique constraint on `event_id` exists\nto make unnecessary (R-12.3). There is no `409` on this route.\n\nThe document is stored verbatim. No `response_model`, no reshaping, no\nvalidation beyond the handful of fields the index sorts and filters on \u2014 a\nfield the device adds must reach the browser, not be quietly dropped on the\nway in, which is the same rule the rollup routes follow.",
    "operationId": "push_event_api_devices_events_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-device-key",
      "required": true,
      "schema": {
       "title": "X-Device-Key",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-device-id",
      "required": true,
      "schema": {
       "title": "X-Device-Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "title": "Payload"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Push Event",
    "tags": [
     "devices"
    ]
   }
  },
...
 }
}